Move use of orders pageviews and their weighting to system configuration.

+++ src/app/code/community/Meanbee/PersonalisedContent/Helper/Customer/Categories.php
@@ -2,12 +2,8 @@
 
 class Meanbee_PersonalisedContent_Helper_Customer_Categories
 {
-    /**
-     * Value importance weighting of products viewed and ordered
-     */
-    const ORDER_SCORE = 50;
-    const VIEW_SCORE = 5;
-
+    /** @var Meanbee_PersonalisedContent_Helper_Config $config */
+    protected $_config;
 
 
     /**
@@ -17,27 +13,36 @@ class Meanbee_PersonalisedContent_Helper_Customer_Categories
      */
     public function reindexCustomerCategories()
     {
+        /** @var Meanbee_PersonalisedContent_Helper_Config $config */
+        $config = $this->getConfig();
+        if (!$config->isEnabled()) {
+            return $this;
+        }
+
         // Empty index table
 
         /** @var Meanbee_PersonalisedContent_Model_Resource_Customer_Categories $customerCategoriesResource */
         $customerCategoriesResource = Mage::getResourceModel('meanbee_personalisedcontent/customer_categories');
         $customerCategoriesResource->truncate();
 
-        // Iterate through all orders from logged in customers.
+        if ($config->isUseOrders()) {
+            // Iterate through all orders from logged in customers.
+            /** @var Mage_Sales_Model_Resource_Order_Collection $sales */
+            $sales = Mage::getModel('sales/order')->getCollection()
+                ->addFieldToFilter('customer_id', array('notnull' => true));
 
-        /** @var Mage_Sales_Model_Resource_Order_Collection $sales */
-        $sales = Mage::getModel('sales/order')->getCollection()
-            ->addFieldToFilter('customer_id', array('notnull' => true));
+            Mage::getSingleton('core/resource_iterator')->walk($sales->getSelect(), array(array($this, 'getProductsForOrder')));
+        }
 
-        Mage::getSingleton('core/resource_iterator')->walk($sales->getSelect(), array(array($this, 'getProductsForOrder')));
+        if ($config->isUsePageviews()) {
+            // Iterate through all views products
+            // Can't use built-in reports very easily as they're very tailored around product collections and admin reports
+            $zendDb = Mage::getModel('core/resource')->getConnection('core_read');
+            $tableName = Mage::getSingleton('core/resource')->getTableName('reports/viewed_product_index');
+            $select = sprintf('SELECT * FROM %s WHERE customer_id is not NULL;', $tableName);
 
-        // Iterate through all views products
-        // Can't use built-in reports very easily as they're very tailored around product collections and admin reports
-        $zendDb = Mage::getModel('core/resource')->getConnection('core_read');
-        $tableName = Mage::getSingleton('core/resource')->getTableName('reports/viewed_product_index');
-        $select = sprintf('SELECT * FROM %s WHERE customer_id is not NULL;', $tableName);
-
-        Mage::getSingleton('core/resource_iterator')->walk($select, array(array($this, 'getProductsForView')), null, $zendDb );
+            Mage::getSingleton('core/resource_iterator')->walk($select, array(array($this, 'getProductsForView')), null, $zendDb);
+        }
 
         return $this;
     }
@@ -88,7 +93,9 @@ class Meanbee_PersonalisedContent_Helper_Customer_Categories
             return;
         }
 
-        $score = self::ORDER_SCORE / count($categoryIds);
+        $config = $this->getConfig();
+
+        $score = $config->getOrderWeighting()/ count($categoryIds);
         foreach ($categoryIds as $categoryId) {
             /** @var Meanbee_PersonalisedContent_Model_Resource_Customer_Categories_Collection $customerCategories */
             $customerCategories = Mage::getModel('meanbee_personalisedcontent/customer_categories')->getCollection()
@@ -136,7 +143,9 @@ class Meanbee_PersonalisedContent_Helper_Customer_Categories
             return;
         }
 
-        $score = self::VIEW_SCORE / count($categoryIds);
+        $config = $this->getConfig();
+
+        $score = $config->getPageviewWeighting() / count($categoryIds);
         foreach ($categoryIds as $categoryId) {
             /** @var Meanbee_PersonalisedContent_Model_Resource_Customer_Categories_Collection $customerCategories */
             $customerCategories = Mage::getModel('meanbee_personalisedcontent/customer_categories')->getCollection()
@@ -160,5 +169,16 @@ class Meanbee_PersonalisedContent_Helper_Customer_Categories
 
             $customerCategory->save();
         }
+    }
+
+    /**
+     * @return Meanbee_PersonalisedContent_Helper_Config
+     */
+    private function getConfig() {
+        if (!$this->_config) {
+            $this->_config = Mage::helper('meanbee_personalisedcontent/config');
+        }
+
+        return $this->_config;
     }
 }
